MFAVA: working in OS space; fixed some things for protected mode, unsure if it works now? MFVS: fixed issues with tnef stuff, one item todo. completed UnZip. ConfigEditor: added virus scan rule editor. MFMail/MFWork: removed some old debugging msgs

+++ MailFilter/Main/MFUnZip.h
@@ -124,7 +124,7 @@ public:
 protected:
 	unzFile zipFile;
 	/* extract the currently selected file */
-	int ExtractCurrentFile(const char* localFilename);
+	long ExtractCurrentFile(const char* localFilename);
 };
 
 #endif	// MFUnZip_H
